requirements.txt
Django==3.1.1
django-cors-headers==3.8.0
django-utils-six==2.0
dj-database-url==0.4.2
gunicorn==20.0.4
psycopg2==2.7.3.2
pytz==2017.3
requests==2.20.1
six==1.11.0
whitenoise==5.2.0
django-allauth==0.45.0

djangorestframework==3.12.4
tensorflow-cpu==2.6.0
numpy==1.19.2
joblib==1.0.1
scipy==1.2.1
librosa==0.8.1
azure-storage-blob==12.8.1